+++ src/GjenbruksElement/Kurver/KurveFunksjoner/makeUrl.js
@@ -1,0 +1,7 @@
+export default function makeUrl(punkt, gradient) {
+  // TODO:
+  if (!punkt) return `https://romlig.artsdatabanken.no/${gradient.url}`;
+  return `https://romlig.artsdatabanken.no/statistikk/grid1d?punkter=${
+    punkt.url
+  }&raster=${gradient.url}`;
+}
