+++ src/GjenbruksElement/Kurver/KurveFunksjoner/prep.js
@@ -1,0 +1,4 @@
+export default function prep(stats, grad) {
+  if (!stats) return grad;
+  return stats.map((y, i) => y / (grad[i] || 1));
+}
