+++ src/GjenbruksElement/Statistikk/Illustrasjon.js
