+++ src/GjenbruksElement/Statistikk/Tekstboks.js
